+++ wallet-core/src/test/java/eu/europa/ec/eudi/wallet/issue/openid4vci/OpenId4VciManagerBuilderTest.kt
@@ -35,6 +35,7 @@ class OpenId4VciManagerBuilderTest {
         useDPoPIfSupported = true,
         parUsage = OpenId4VciManager.Config.ParUsage.IF_SUPPORTED,
         proofTypes = listOf(OpenId4VciManager.Config.ProofType.JWT),
+        debugLogging = true,
     )
 
     @Test