+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.9.5-SNAPSHOT]
+
+__12 Jun 2024__
+
+- Bump eudi-lib-jvm-openid4vci-kt to v0.3.1
+- Configurable debug logging for OpenId4VCI
+
 ## [0.9.4-SNAPSHOT]
 
 __10 Jun 2024__